Improve types and interfaces.ts for improved readability
