+++ kpm/kpm-backend-interface/src/errors.ts
@@ -1,5 +1,23 @@
-export type TAuthErrorType = "bla" | "bli";
+/**
+ * General Error Types
+ */
 
+export type TGotErrType = "NotAvailable" | "BadResponse" | "TimeoutError";
+
+export type TCorsError = "InvalidOrigin";
+
+export type APIAuthErrType = "ClientResponseError" | "TypeError";
+
+export type APIMutedAuthErrType =
+  | "LoginRequired"
+  | "SessionExpired"
+  | "SessionStoreError"
+  | "AuthServiceMiscError"
+  | "NoSessionUser";
+
+/**
+ * API Error Payload Type
+ */
 export type TApiError = {
   statusCode: number;
   name?: string;
